internal/e2e: consolidate e2e helpers
