--- internal/e2e/e2e_test.go
@@ -1,46 +1,0 @@
-// Copyright © 2018 Heptio
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// End to ends tests for translator to grpc operations.
-package e2e
-
-import (
-	"testing"
-
-	"github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	"github.com/gogo/protobuf/proto"
-	"github.com/gogo/protobuf/types"
-)
-
-func any(t *testing.T, pb proto.Message) types.Any {
-	t.Helper()
-	any, err := types.MarshalAny(pb)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return *any
-}
-
-func assertEqual(t *testing.T, want, got *v2.DiscoveryResponse) {
-	t.Helper()
-	m := proto.TextMarshaler{Compact: true, ExpandAny: true}
-	a := m.Text(want)
-	b := m.Text(got)
-	if a != b {
-		m := proto.TextMarshaler{
-			Compact:   false,
-			ExpandAny: true,
-		}
-		t.Fatalf("\nexpected:\n%v\ngot:\n%v", m.Text(want), m.Text(got))
-	}
-}
